requirements.txt
pyserial>=3.5
WMI>=1.5.1
pypiwin32>=223
requests>=2.28.0
python-dotenv>=0.20.0
pyyaml>=6.0
